Update the server to take in group & port numbers from client.
The client executable takes group ID and port numbers as command
line arguments and sends them to server, now when a client sends
a message to a server, server sends this messages to all the clients
with group ID equal to that of the client which sent the message.
Essentially, broadcasting the message across the group.

Now, the two way chat just becomes a group with two people.