Working version of voice call app.
A user needs to enter as a group 2 member to listen in to the
conversations of others.

So, in a voice call, if a person A wants to talk to another B, the
other person (B) first enters as a member of group 2, starts listening
to what A says (he should be in group other than 2). Now if A wants
 to talk, he should restart as a group 2 member and B should join
 again as member from group other than 2.